Attempt to find a vanity name when composing message
When we compose the message, we query our experimental DNS zone for a
vanity name. This allows the bot to tweet the vanity name of the Baker.

If you wish to have the bot tweet a vanity name for your address when
you vote, then please send an email to frontdesk@ecadlabs.com